zack/r10k has been moved to puppet/r10k
Setting up a new puppet server today, I discovered that zack/r10k was marked as deprecated/EOL (just yesterday!) and has been transferred to the Vox Pupuli "puppet" namespace. It also got some nice updates... so I installed the new one instead.

Then I tried installing your puppet module, and of course it failed because of the dependency conflict. Comparing the code between the old and new versions of the r10K module there are only a few differences, and nothing that appears to be backwards-incompatible in any way.